[wip-lw] new game
The most significant changes are in the `GameState.cs` and `MinesweeperFace.razor.cs` files. In `GameState.cs`, the `GameModeInstance` property now has a default value, and the `CellStates` property is now nullable with a private set. A new `NewGame` method has been added, which sets the `GameModeInstance` property and calls the parameterless `NewGame` method. The `SetMines` method now throws an `InvalidOperationException` if `CellStates` is `null`. In `MinesweeperFace.razor.cs`, a new private variable `_isFocus` has been added, and the `HandleMouseDown`, `HandleMouseUp`, and `HandleMouseOut` events have been updated. A new `GameState` parameter has been added to the `MinesweeperFace` class.

Changes:

1. In `CellState.cs`, the properties `IsMine`, `IsFlagged`, and `IsRevealed` now have a default value of `false`.
2. In `MinesweeperFace.razor.cs`, a new private variable `_isFocus` has been added. This variable is set to `true` when the `HandleMouseDown` event is triggered. The `HandleMouseUp` and `HandleMouseOut` events now only change the `Face` property if `_isFocus` is `true`.
3. A new `GameState` parameter has been added to the `MinesweeperFace` class in `MinesweeperFace.razor.cs`.
4. In `GameState.cs`, the `GameModeInstance` property now has a default value of `default!`. The `CellStates` property is now nullable and private set.
5. A new `NewGame` method has been added to `GameState.cs` that takes an `IGameMode` parameter. This method sets the `GameModeInstance` property and calls the parameterless `NewGame` method.
6. The parameterless `NewGame` method in `GameState.cs` now calls the `Initialize` method and sets the `GameStatus` property to `AwaitingFirstMove`. The `Initialize` method no longer sets the `GameStatus` property.
7. In the `SetMines` method of the `GameState` class in `GameState.cs`, a check has been added to throw an `InvalidOperationException` if `CellStates` is `null`.